1_pytest_e2e/requirements.txt
pytest
selenium
black
flake8
isort
Appium-Python-Client>=0.28
SauceClient